\name{ResourceSelection-package}
\alias{ResourceSelection-package}
\alias{ResourceSelection}
\docType{package}
\title{
Resource Selection (Probability) Functions for Use-Availability Data
}
\description{
  Resource Selection (Probability) Functions for 
  use-availability wildlife data as described in 
  Lele and Keim (2006, Ecology 87, 3021--3028),
  and Lele (2009, J. Wildlife Management 73, 122--127).
}
\details{
\code{\link{rsf}}: Resource Selection Functions (RSF)

\code{\link{rspf}}: Resource Selection Probability Functions (RSPF)
}
\author{
Subhash R. Lele, Jonah L. Keim, Peter Solymos

Maintainer: Peter Solymos <solymos@ualberta.ca>
}
\references{
Lele, S.R. (2009) 
A new method for estimation of resource selection probability function. 
Journal of Wildlife Management 73, 122--127.

Lele, S. R. &  Keim, J. L. (2006) 
Weighted distributions and estimation of resource selection probability functions. 
Ecology 87, 3021--3028.
}
\keyword{ package }
\seealso{
\code{\link{rsf}}, \code{\link{rspf}}, \code{\link{kdepairs}}, \code{\link{hoslem.test}}
}
