\name{ResourceSelection-package}
\alias{ResourceSelection-package}
\alias{ResourceSelection}
\docType{package}
\title{
Resource Selection (Probability) Functions for Use-Availability Data
}
\description{
Resource Selection (Probability) Functions
for use-availability wildlife data
based on weighted distributions as described in
Lele and Keim (2006), Lele (2009), and Solymos & Lele (2016).
}
\details{
\code{\link{rsf}}: Resource Selection Functions (RSF)

\code{\link{rspf}}: Resource Selection Probability Functions (RSPF)

\code{\link{hoslem.test}}: Hosmer-Lemeshow Goodness of Fit (GOF) Test

Visual summaries: \code{\link{kdepairs}} for 2D scatterplots and
\code{\link{mep}} for marginal effect plots based on fitted model objects.
}
\author{
Subhash R. Lele, Jonah L. Keim, Peter Solymos

Maintainer: Peter Solymos <solymos@ualberta.ca>
}
\references{
Lele, S.R. (2009)
A new method for estimation of resource selection probability function.
Journal of Wildlife Management 73, 122--127.
<doi:10.2193/2007-535>

Lele, S. R. &  Keim, J. L. (2006)
Weighted distributions and estimation of resource selection probability function.
Ecology 87, 3021--3028.
<doi:10.1890/0012-9658(2006)87%5B3021:WDAEOR%5D2.0.CO;2>

Solymos, P. & Lele, S. R. (2016)
Revisiting resource selection probability functions and single-visit methods:
clarification and extensions.
Methods in Ecology and Evolution 7, 196--205.
<doi:10.1111/2041-210X.12432>
}
\keyword{ package }
\seealso{
\code{\link{rsf}}, \code{\link{rspf}}, \code{\link{kdepairs}}, \code{\link{mep}},
\code{\link{hoslem.test}}
}
